fix: parse custom args of Node to get correct non default ports
If the custom args where changing the zmq/rpc ports of the Node,
Gupaxx would not have the right nodes to share with other processes nor
could fetch the data from the Node (Unknown status).

There was also the same issue if a non default port was set in advanced submenu but
the Node was started in Simple mode.