feat: renaming gupaxx to gupax
updated the README to not focus on XvB anymore but presenting Gupax globally in a simple way.
Some portion of the original README of gupax will be moved to documentation.

Gupax v2 will check if gupaxx was used more recently than gupax and will apply the settings from it
if gupax last use wasn't v2.

fix